Revert "create cube visualization"
